issue #26 resolved (someone is hoolding stale ingame context)

+++ src/net/torvald/terrarum/gameactors/ActorWBMovable.kt
@@ -345,7 +345,7 @@ open class ActorWBMovable(renderOrder: RenderOrder, val immobileBody: Boolean = 
     val feetPosTile: Point2i = Point2i(0,0)
         //get() = Point2i(hIntTilewiseHitbox.centeredX.floorInt(), hIntTilewiseHitbox.endY.floorInt())
 
-    override fun run() = update(AppLoader.UPDATE_RATE.toFloat())
+    override fun run() = update(AppLoader.UPDATE_RATE)
 
     /**
      * Add vector value to the velocity, in the time unit of single frame.
