+++ src/net/torvald/terrarum/modulebasegame/gameactors/ThreadActorUpdate.kt
@@ -11,13 +11,13 @@ class ThreadActorUpdate(val startIndex: Int, val endIndex: Int) : Runnable {
     override fun run() {
         for (i in startIndex..endIndex) {
             val it = Terrarum.ingame!!.actorContainerActive[i]
-            it.update(AppLoader.UPDATE_RATE.toFloat())
+            it.update(AppLoader.UPDATE_RATE)
 
             if (it is Pocketed) {
                 it.inventory.forEach { inventoryEntry ->
-                    ItemCodex[inventoryEntry.item]?.effectWhileInPocket(AppLoader.UPDATE_RATE.toFloat())
+                    ItemCodex[inventoryEntry.item]?.effectWhileInPocket(AppLoader.UPDATE_RATE)
                     if (it.equipped(inventoryEntry.item)) {
-                        ItemCodex[inventoryEntry.item]?.effectWhenEquipped(AppLoader.UPDATE_RATE.toFloat())
+                        ItemCodex[inventoryEntry.item]?.effectWhenEquipped(AppLoader.UPDATE_RATE)
                     }
                 }
             }
